sstablelib.py: Stream: add instantiate()
Converts a type that has one or more member types to a canonical one
that only takes a stream instance as its only parameter.
This is useful for creating pre-instantiated templated types, e.g.:

    instantiate(array32, int64)

To create an array of ints. The type created this way just needs a
`stream` instance and will return the parsed array when called.