sstablelib.py: add parse()
Allow parsing data based on a schema. Schema is a tuple of key-value
pairs. The key is a name, while the value is a type.
The result is a dictionary with with the passed in keys and
corresponding values parsed with the respective type.
Example:

    schema = (
        ('field1', Stream.uint64),
        ('field2', Stream.string16),
        ('field3', Stream.instantiate(Stream.array32, Stream.int32),
    )
    parse(stream, schema)

An example result could be:

    {
        'field1': 10,
        'field2': "some value",
        'field3': [2134, 21, 987],
    }

`parse()` allows using nice declarative schemas to describe complex data
structures, resulting in much more human readable code than an imperative
equivalent.