sstablelib.py: Stream.string16(): don't throw on undecodable strings
Some statistics files will contain strings that are not decodable with
UTF-8 for reasons that are beyond me. I have spent enough time on trying
to figure this out, for now in this case just return the raw bytes. We
can revisit this later when one of those strings becomes important (I
left a FIXME).