Move reusable code into the `sstable_tools` python module
The long-term vision is that all logic is in a reusable form inside that
module, so other code can be built on it. Top level scripts should just
be one users of the module. For now only the `sstable-statistics.py`
code is moved into the module, along with the `sstablelib.py` utility
code. The other scripts could not be trivially converted into a reusable
form so we'll defer that for now.